requirements.txt
aider-chat
black
click
jinja2
Pillow
pydantic
pytest
requests
isort